Update pandas.html
A student noticed that section 9.2 of the book (that describes auto sorting of indices when creating a series from a labeled dictionary) conflicts with the video in Module 5.1.2. This is due to a change in the pandas library for python 3.6+ that no longer auto sorts. Release note here: https://pandas.pydata.org/docs/dev/whatsnew/v0.23.0.html#instantiation-from-dicts-preserves-dict-insertion-order-for-python-3-6

Book text revised to show the properly non-sorted series output as well as remove the paragraph that describes auto-sorting.