Update pandas.html
At the end of Section 9.3.3, the book discusses the _df.loc method of index label retrieval. However, based on the examples just above, it is not obvious that the index must be reset before you can reference 'Ada' or 'Bob' as is shown in the examples.

Suggest adding this paragraph (or something similar) to clarify.